+++ src/event.h
@@ -37,9 +37,9 @@ bool event_accept_submit(Event*, struct io_uring*, fd socket,
                          struct sockaddr_in* out_client_addr,
                          socklen_t*          inout_addr_len);
 bool event_send_submit(Event*, struct io_uring*, fd socket, CString data,
-                       unsigned sqe_flags);
+                       uint8_t sqe_flags);
 bool event_recv_submit(Event*, struct io_uring*, fd socket, String out_data);
 bool event_read_submit(Event*, struct io_uring*, fd file, String out_data,
-                       size_t nbytes, off_t offset, unsigned sqe_flags);
+                       size_t nbytes, off_t offset, uint8_t sqe_flags);
 bool event_close_submit(Event*, struct io_uring*, fd socket);
 bool event_timeout_submit(Event*, struct io_uring*, time_t sec, time_t nsec);
